+++ dockstore-webservice/src/main/java/io/dockstore/webservice/resources/EntryResource.java
@@ -21,14 +21,12 @@ import static io.dockstore.webservice.resources.ResourceConstants.JWT_SECURITY_D
 
 import com.codahale.metrics.annotation.Timed;
 import io.dockstore.common.DescriptorLanguage;
-import io.dockstore.common.SourceControl;
 import io.dockstore.webservice.CustomWebApplicationException;
 import io.dockstore.webservice.DockstoreWebserviceConfiguration;
 import io.dockstore.webservice.api.SyncStatus;
 import io.dockstore.webservice.core.Category;
 import io.dockstore.webservice.core.CollectionOrganization;
 import io.dockstore.webservice.core.DescriptionMetrics;
-import io.dockstore.webservice.core.Doi;
 import io.dockstore.webservice.core.Entry;
 import io.dockstore.webservice.core.LambdaEvent;
 import io.dockstore.webservice.core.OrcidPutCode;
@@ -40,14 +38,12 @@ import io.dockstore.webservice.core.User;
 import io.dockstore.webservice.core.Version;
 import io.dockstore.webservice.core.Workflow;
 import io.dockstore.webservice.core.WorkflowMode;
-import io.dockstore.webservice.core.WorkflowVersion;
 import io.dockstore.webservice.core.database.VersionVerifiedPlatform;
 import io.dockstore.webservice.helpers.LambdaUrlChecker;
 import io.dockstore.webservice.helpers.ORCIDHelper;
 import io.dockstore.webservice.helpers.PublicStateManager;
 import io.dockstore.webservice.helpers.StateManagerMode;
 import io.dockstore.webservice.helpers.TransactionHelper;
-import io.dockstore.webservice.helpers.ZenodoHelper;
 import io.dockstore.webservice.jdbi.EntryDAO;
 import io.dockstore.webservice.jdbi.EventDAO;
 import io.dockstore.webservice.jdbi.LambdaEventDAO;
@@ -100,12 +96,8 @@ import java.net.HttpURLConnection;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.http.HttpResponse;
-import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
 import java.util.function.BiFunction;
@@ -733,93 +725,6 @@ public class EntryResource implements AuthenticatedResourceInterface, AliasableR
 
     private boolean isPrimaryDescriptor(String path, SourceFile sourceFile) {
         return sourceFile.getPath().equals(path);
-    }
-
-    /**
-     * Scans Zenodo for DOIs issued against GitHub repos with registered entries in Dockstore, updating the Dockstore entries
-     * with those DOIs.
-     *
-     * @param user
-     * @param filter - optional filter to scope to a single GitHub repository in the format myorg/myrepo
-     * @return
-     */
-    @POST
-    @RolesAllowed({"admin", "curator"})
-    @Path("/updateDois")
-    @UnitOfWork
-    @Timed
-    @Operation(operationId = "updateDois", description = "Searches Zenodo for DOIs referencing GitHub repos, and updates Dockstore entries with them", security = @SecurityRequirement(name = JWT_SECURITY_DEFINITION_NAME))
-    public List<Workflow> updateDois(@Parameter(hidden = true) @Auth User user,
-            @Parameter(description = "Optional GitHub full repository name, e.g., myorg/myrepo, to only update entries for that repository") @QueryParam("filter") String filter) {
-        final List<Workflow> publishedWorkflows = getPublishedGitHubWorkflows(filter);
-
-        // Get a map of GitHub repos to Dockstore workflows
-        final Map<String, List<Workflow>> repoToWorkflowsMap = publishedWorkflows.stream()
-                .filter(w -> w.getSourceControl() == SourceControl.GITHUB)
-                .collect(Collectors.groupingBy(w -> w.getOrganization() + '/' + w.getRepository()));
-
-        final List<ZenodoHelper.GitHubRepoDois> gitHubRepoDois = repoToWorkflowsMap.keySet().stream().sorted()
-                .map(ZenodoHelper::findDoisForGitHubRepo)
-                .flatMap(Collection::stream)
-                .toList();
-
-        final Set<Workflow> updatedWorkflows = new HashSet<>();
-        gitHubRepoDois.stream().forEach(gitHubRepoDoi -> {
-            final String repo = gitHubRepoDoi.repo();
-            final String conceptDoi = gitHubRepoDoi.conceptDoi();
-            final List<Workflow> workflows = repoToWorkflowsMap.get(repo);
-            workflows.stream().forEach(workflow -> {
-                final Doi existingGitHubDoi = workflow.getConceptDois().get(Doi.DoiInitiator.GITHUB);
-                if (existingGitHubDoi != null && conceptDoi.equals(existingGitHubDoi.getName())) {
-                    LOG.warn("Skipping DOI %s for workflow %s because it already has a Zenodo DOI from GitHub".formatted(conceptDoi, workflow.getWorkflowPath()));
-                } else {
-                    final boolean noDoiToStart = workflow.getConceptDois().isEmpty();
-                    if (existingGitHubDoi == null) { // No Concept DOI yet, add it.
-                        workflow.getConceptDois().put(Doi.DoiInitiator.GITHUB,
-                                ZenodoHelper.getDoiFromDatabase(Doi.DoiType.CONCEPT, Doi.DoiInitiator.GITHUB, conceptDoi));
-                        updatedWorkflows.add(workflow);
-                    }
-                    // Add version DOI(s) to new or existing Concept DOI.
-                    gitHubRepoDoi.tagAndDoi().stream().forEach(tagAndDoi -> {
-                        final WorkflowVersion workflowVersion = workflowVersionDAO.getWorkflowVersionByWorkflowIdAndVersionName(
-                                workflow.getId(), tagAndDoi.gitHubTag());
-                        if (workflowVersion != null) {
-                            if (workflowVersion.getDois().get(Doi.DoiInitiator.GITHUB) == null) {
-                                final Doi versionDoi = ZenodoHelper.getDoiFromDatabase(Doi.DoiType.VERSION, Doi.DoiInitiator.GITHUB,
-                                        tagAndDoi.doi());
-                                workflowVersion.getDois().put(Doi.DoiInitiator.GITHUB, versionDoi);
-                                updatedWorkflows.add(workflow);
-                            }
-                        }
-                    });
-                    if (noDoiToStart) {
-                        // The default DOI selection is USER. If there was no DOI to begin with, set it to GITHUB so it will show up in the UI.
-                        workflow.setDoiSelection(Doi.DoiInitiator.GITHUB);
-                    }
-                }
-            });
-        });
-        return updatedWorkflows.stream().toList();
-    }
-
-    /**
-     * Returns a list of published GitHub workflows. If <code>optionalFilter</code> is set, then only returns workflows for that GitHub
-     * organization and repository. Throws a CustomWebApplicationException if <code>optionalFilter</code> is set, and its format is not
-     * <code>[organization]/[repository]</code>.
-     * @param optionalFilter a filter in the format "organization/repository", or null/empty
-     * @return
-     */
-    private List<Workflow> getPublishedGitHubWorkflows(String optionalFilter) {
-        if (StringUtils.isNotBlank(optionalFilter)) {
-            final String[] split = optionalFilter.split("/");
-            if (split.length != 2) {
-                throw new CustomWebApplicationException("Filter '%s' must be of the format org/repo".formatted(optionalFilter), HttpStatus.SC_BAD_REQUEST);
-            }
-            final String org = split[0];
-            final String repository = split[1];
-            return workflowDAO.findByOrganizationAndRepository(SourceControl.GITHUB, org, repository);
-        }
-        return workflowDAO.findAllPublished(0, Integer.MAX_VALUE, null, null, null);
     }
 
     /**
